+++ src/main/java/development/v/development/domain/repositories/UserRepository.java
@@ -5,5 +5,5 @@ import java.util.Optional;
 import development.v.development.domain.models.User;
 
 public interface UserRepository {
-    Optional<User> findById(Integer usuId);
+    Optional<User> findById(Long usuId);
 }
